Refactor ConnectorComparator into mixin modules
Split the 4108-line ConnectorComparator class into a comparator/ package of
concern-focused mixins, leaving connector_comparator.py as the entry point.

- comparator/template_resolver.py: TemplateResolverMixin
- comparator/config_mapper.py: ConfigMapperMixin
- comparator/config_deriver.py: ConfigDeriverMixin

Pure structural refactor: all 87 methods moved verbatim, no logic changes.
The public API (from connector_comparator import ConnectorComparator, class
constants, static parse_connector_file) is unchanged.
