+++ src/comparator/__init__.py
@@ -1,0 +1,1 @@
+"""ConnectorComparator mixins (template resolution, config mapping, config derivation)."""
